Delete the test database before all integration tests run
