+++ tests/setupJest.ts
@@ -1,4 +1,14 @@
-export {};
+import path from "path";
+import { fsUnlink } from "./fsUtils";
 
 // 30-second timeout for E2E tests
 jest.setTimeout(30000);
+
+async function deleteTestDatabase() {
+  const dbDir = path.resolve(__dirname, "./db/db.sqlite");
+  await fsUnlink(dbDir);
+}
+
+beforeAll(async () => {
+  await deleteTestDatabase();
+});
